cmd/gxz: Introduce config creator functions
The condfigurations for xz and lzma reader and writers are currently
created inside the decompressor and compressor functions for each
compression format. This commit extracts the functionality to specific
function for each required configuration. Which will allow us to
implement more complex logic to for the creation of the configurations
later.