Introduce NewWriterCfg and NewReaderCfg to package APIs
The old API used methods on the Config structures to create new readers
and writers with configuration parameters. This is now changed to
classic NewReaderCfg or NewWriterCfg functions. This is simpler and also
better represented in godoc since all functions generating a reader or
writer are in one place.

Additionally the verify method has been removed from the APIs.