lzma: Fix issues with hash chain finder
The hash chain finder used complex double chained links and broke. This
implementation has gone back to a simpler approach. It doesn't store the
hashes in the table and recomputes them for the resizing and implement
the hash chain with distance deltas helping us to identify entries that
are too old. The test cases appear to run.